+++ CHANGELOG.rst
@@ -12,7 +12,7 @@ v1.2.0
 
    | If it's false, the WebElement is searched whenever is needed (default value)
    | If it's true, the WebElement is saved in PageElement to avoid searching for the same element multiple times. Useful
-   |  in mobile testing when searching for an element can take a long time.
+   | in mobile testing when searching for an element can take a long time.
 - Refactored reset_object method. Now it has an optional parameter with the driver_wrapper.
 - Fix error downloading videos after failed tests
 - System property 'Config_environment' is used to select config files, e.g., to read android-properties.cfg file:
@@ -21,12 +21,14 @@ v1.2.0
 
     $ Config_environment=android nosetests web/tests/test_web.py
 
-Behave user property 'env' is deprecated, use 'Config_environment' instead:
+- Behave user property 'env' is deprecated, use 'Config_environment' instead:
 
 .. code:: console
 
     $ behave -D Config_environment=android
 - Fix error in visual tests when excluding elements in a scrolled page
+- New config property 'restart_driver_fail' in [Driver] section to restart the driver when the test fails even though
+  the value of *reuse_driver* property is *true*
 
 v1.1.3
 ------